+++ jsapp/js/components/reports/reportColorsEditor.component.tsx
@@ -3,30 +3,29 @@ import bem from 'js/bem';
 import {CHART_COLOR_SETS} from './reportsConstants';
 import type {ReportStyle, ChartColorSet} from './reportsConstants';
 
-interface ChartColorsPickerProps {
-  onChange: (params: {default: boolean}, value: {report_colors: string[]}) => void;
-  defaultStyle: ReportStyle;
+interface ReportColorsEditorProps {
+  onChange: (newColors: string[]) => void;
+  /** The style that is being edited */
+  style: ReportStyle;
 }
 
-export default function ChartColorsPicker(props: ChartColorsPickerProps) {
-  function defaultReportColorsChange(value: number) {
+export default function ReportColorsEditor(props: ReportColorsEditorProps) {
+  function onStyleChange(value: number) {
     let newColors = CHART_COLOR_SETS[0].colors;
     if (CHART_COLOR_SETS[value]?.colors) {
       newColors = CHART_COLOR_SETS[value].colors;
     }
 
-    props.onChange({default: true}, {report_colors: newColors});
+    props.onChange(newColors);
   }
 
-  // Not sure why this is called "is default", and not simply "is active". This
-  // needs some more investigation.
-  function isDefaultValue(set: ChartColorSet, index: number) {
-    if (props.defaultStyle.report_colors === undefined && index === 0) {
-      return true;
-    }
-
+  /**
+   * Checks if given set of colors is the one that is being used by the style
+   * that is being edited in this component instance.
+   */
+  function isChecked(set: ChartColorSet) {
     return (
-      JSON.stringify(props.defaultStyle.report_colors) ===
+      JSON.stringify(props.style.report_colors) ===
       JSON.stringify(set.colors)
     );
   }
@@ -39,8 +38,8 @@ export default function ChartColorsPicker(props: ChartColorsPickerProps) {
             type='radio'
             name='chart_colors'
             value={index}
-            checked={isDefaultValue(set, index)}
-            onChange={() => defaultReportColorsChange(index)}
+            checked={isChecked(set)}
+            onChange={() => onStyleChange(index)}
             id={'type-' + set.label}
           />
